==> 1_exclusion.Rmd <==
# Reasons for Exclusion 

```{r}
source("0_helpers.R")
library(tidylog)
load("data/cleaned.rdata")

library(knitr)
opts_chunk$set(fig.width = 9, fig.height = 7, cache = T, warning = T, message = T, cache = F, error = FALSE)

comma_separated_to_columns <- function(df, col) {
  colname <- deparse(substitute(col))
  df$splitcol <- df %>% pull(colname)
  separate_rows(df, splitcol, convert = TRUE, sep = ", ") %>% 
    mutate(splitcol = if_else(is.na(splitcol), "no", 
                        if_else(splitcol == "" | 
                                  splitcol %in% c(), "included", as.character(splitcol)))) %>% 
    mutate(#splitcol = stringr::str_c(colname, "_", splitcol), 
           value = 1) %>% 
    spread(splitcol, value, fill = 0) %>% 
    select(-colname)
}

all_survey_length <- nrow(all_surveys)
diary_length <- nrow(diary)
diary_social_length <- nrow(diary_social)
```

```{r}
diary$reasons_for_exclusion_diary <- ""
```


Lab participants that did not do the online diary (not merged)

```{r}
diary <- diary %>% 
  mutate(reasons_for_exclusion_diary = str_c(reasons_for_exclusion_diary,
                                            if_else(is.na(session), "lab_only, ", "", "")
                                       )
  )
```

Disclosed that they responded dishonestly on that day.

```{r}
diary <- diary %>% 
  mutate(reasons_for_exclusion_diary = str_c(reasons_for_exclusion_diary,
                                            if_else(dishonest_discard == 1, "dishonest_answer, ", "", "")
                                       )
  )
```

Did not finish diary entry.

```{r}
diary <- diary %>% 
  mutate(reasons_for_exclusion_diary = str_c(reasons_for_exclusion_diary,
                                            if_else(is.na(ended_diary) & !is.na(modified_diary), "did_not_finish_entry, ", "", "")
                                       )
  )
```


Cycle shorter than 20 days.
```{r}
diary <- diary %>% 
  mutate(reasons_for_exclusion_diary = str_c(reasons_for_exclusion_diary,
                                            if_else(coalesce(minimum_cycle_length_diary, as.numeric(menstruation_length)) < 20, "cycle_shorter_than_20, ", "", "")
                                       )
  )
```

Cycle longer than 40 days.
```{r}
diary <- diary %>% 
  mutate(reasons_for_exclusion_diary = str_c(reasons_for_exclusion_diary,
                                            if_else(coalesce(minimum_cycle_length_diary, as.numeric(menstruation_length)) > 40, "cycle_longer_than_40, ", "", "")
                                       )
  )
```

Next menstrual onset not observed

```{r}
diary <- diary %>% 
  mutate(reasons_for_exclusion_diary = str_c(reasons_for_exclusion_diary,
                                            if_else(
                                              menstruation_regular == 1 &
                                              coalesce(minimum_cycle_length_diary, as.numeric(menstruation_length)) <= 40 &
                                              coalesce(minimum_cycle_length_diary, as.numeric(menstruation_length)) >= 20 &
                                              is.na(fertile_fab), "next_menstrual_onset_unobserved, ", "", "")
                                       )
  )
```


Skipped this diary day (days after dropping out not included)
```{r}
diary <- diary %>% 
  mutate(reasons_for_exclusion_diary = str_c(reasons_for_exclusion_diary,
                                            if_else(is.na(ended_diary) & is.na(modified_diary), "skipped_diary_entry, ", "")
                                       )
  )
```

```{r}
usable_diary_days <- diary %>% group_by(session) %>% 
  summarise(usable_diary_days = any(reasons_for_exclusion_diary == ""))
```



## Who completed what?

We create a character variable `reasons_for_exclusion`. We will concatenate (abbreviated)
reasons for exclusion in this variable.

```{r}
all_surveys$reasons_for_exclusion <- ""
```


Did not finish demographics survey
```{r}
all_surveys <- all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(is.na(ended_demo), "didnt_finish_demographics, ", "")
                                       )
  )
```


Reported no regular menstruation
```{r}
all_surveys <- all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(menstruation_regular == 0, "no_regular_menstruation, ", "", "")
                                       )
  )
```


Menopausal or in climacteric period
```{r}
all_surveys <- all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(menopause_yes == 1 | menopause_yes == 2, "menopausal, ", "", "")
                                       )
  )
```

Older than 50
```{r}
all_surveys <- all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(age >= 50, "older_than_50, ", "", "")
                                       )
  )
```

Pregnant
```{r}
all_surveys <- all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(pregnant == 1, "pregnant, ", "", "")
                                       )
  )
```

Breast-feeding
```{r}
all_surveys <- all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(breast_feeding == 1, "breast_feeding, ", "", "")
                                       )
  )
```


Not primarily heterosexual. This excludes women who reported being equally interested in men and women, women who reported being asexual or aromantic, and participants who did not identify as female gender.
```{r}
all_surveys <- all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(sex_orientation >= 4 | gender != 1, "not_heterosexual_female, ", "", "")
                                       )
  )
```

Did not finish personality survey
```{r}
all_surveys <- all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(!is.na(ended_demo) & is.na(ended_initial), "didnt_finish_personality, ", "")
                                       )
  )
```

Changed contraception
```{r}
all_surveys <- all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(change_to_nonhormonal != 0 | change_to_hormonal_contraception != 0, "switched_contraception, ", "", "")
                                       )
  )

```


Taking sex hormones (other than the pill)
```{r}
all_surveys <- all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(medication_name %contains% "Cycloprognova" |
                                                      medication_name %contains% "Cyproderm" |
                                                       medication_name %contains% "DHEA" |
                                                       medication_name %contains% "Hormone" |
                                                       medication_name %contains% "Cyclo-Progynova" |
                                                       medication_name %contains% "Femoston" |
                                                       medication_name %contains% "Gynokadin", "sex_hormones, ", "", "")
                                       )
  )

```

No diary days
```{r}
all_surveys <- all_surveys %>% 
  left_join(diary %>% group_by(session) %>% summarise(diary_days = n_nonmissing(ended_diary)), by = 'session') %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(!is.na(ended_initial) & diary_days == 0, "didnt_do_diary, ", "", "didnt_do_diary, ")
                                       )
  )
```

Fertility never estimable
```{r}
all_surveys <- all_surveys %>% 
  left_join(diary %>% group_by(session) %>% summarise(fertility_days = n_nonmissing(fertile_fab)), by = 'session') %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(menstruation_regular != 0  & diary_days > 0 & fertility_days == 0, 
                                                    "fertility_never_estimable, ", "", "")
                                       )
  )
```


```{r}
all_surveys <- all_surveys %>% left_join(usable_diary_days, by = "session") %>% 
    mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
            if_else(menstruation_regular != 0  & diary_days > 0 & fertility_days > 0 & !usable_diary_days, "diary_days_not_usable, ", "", "")
                                       )
  )
```

Women who are in monogamous heterosexual relationships
```{r}
all_surveys <- all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(hetero_relationship == 0 & relationship_status != 1, "non_heterosexual_relationship, ", "", "")
                                       )
  )
```

```{r}
testthat::expect_equal(all_surveys %>% group_by(session) %>% filter(n() > 1) %>% nrow(),
                       0)
table(all_surveys$reasons_for_exclusion == "", exclude = NULL)
diary <- diary %>% left_join(all_surveys %>% select(session, reasons_for_exclusion), by = 'session') %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion, reasons_for_exclusion_diary))

library(UpSetR)
exclusion_reasons <- all_surveys %>% 
  mutate(reasons_for_exclusion = str_sub(reasons_for_exclusion, 1, -3)) %>% 
  select(session, reasons_for_exclusion) %>% 
  comma_separated_to_columns(reasons_for_exclusion) %>% 
  select(-session)

exclusion_reasons_table <- exclusion_reasons %>% 
  summarise_all(sum) %>% sort() %>% 
  gather(reason, n) %>% 
  left_join(all_surveys %>% mutate(reason = str_sub(reasons_for_exclusion, 1, -3)) %>% group_by(reason) %>% summarise(unique = n())) %>% 
  mutate(unique = if_else(is.na(unique), 0L, unique))

exclusion_reasons_hc <- all_surveys %>% 
  mutate(reasons_for_exclusion = str_sub(reasons_for_exclusion, 1, -3)) %>% 
  select(session, hormonal_contraception, reasons_for_exclusion) %>% 
  mutate(hormonal_contraception = if_else(hormonal_contraception, 1, 0)) %>% 
  comma_separated_to_columns(reasons_for_exclusion) %>% 
  # filter(included == 0) %>%
  select(-session)

exclusion_reasons_hc_table <- exclusion_reasons_hc %>% group_by(hormonal_contraception) %>% 
  summarise_all(sum) %>% gather(reason, n, -hormonal_contraception) %>% 
  spread(hormonal_contraception, n) %>% 
  arrange(`0`)

exclusion_reasons_diary <- diary %>% 
  mutate(reasons_for_exclusion = str_sub(reasons_for_exclusion, 1, -3)) %>% 
  select(session, created_date, reasons_for_exclusion) %>% 
  drop_na(session, created_date) %>% 
  comma_separated_to_columns(reasons_for_exclusion) %>% 
  select( -created_date)

exclusion_reasons_diary_table <- exclusion_reasons_diary %>% 
  select(-session) %>% 
  summarise_all(sum) %>% 
  sort() %>% 
  gather(reason, n) %>% 
  left_join(diary %>% mutate(reason = str_sub(reasons_for_exclusion, 1, -3)) %>% group_by(reason) %>% summarise(unique = n())) %>% 
  mutate(unique = if_else(is.na(unique), 0L, unique)) %>% 
  left_join(exclusion_reasons_diary %>% 
  gather(reason, n, -session) %>% 
  filter(n > 0) %>% 
  distinct(session, reason, n) %>% 
  group_by(reason) %>%
  summarise(n_women = sum(n)))
```

## Reasons for exclusion
```{r fig.cap="How to read this plot: The horizontal green bars show for how many women this reason for exclusion applies. The blue bars show how many women are excluded for multiple reasons (e.g., they're menopausal _and_ not heterosexual). If reasons for exclusion necessarily depend on another (i.e. participants had to finish the first survey to get to the second), we counted only those who had not yet been excluded earlier. If reasons for exclusion depended on each other only stochastically (e.g., age and menopause), we did not do this."}
exclusion_reasons %>% 
  filter(included == 0) %>% 
  select(-included) %>% 
  as.data.frame() %>% 
  {
  upset(., ncol(.), 20, show.numbers = TRUE, order.by = "freq",
      main.bar.color = "#6E8691",
      matrix.color = "#6E8691",
      sets.bar.color = "#53AC9B")
  }
```

```{r}
exclusion_reasons_table %>% 
  knitr::kable(caption = "Reasons for exclusion. _n_ shows the number of affected women, _unique_ those who for whom this was the only reason to be excluded.")
```


### By contraception
```{r}
exclusion_reasons_hc %>% 
  as.data.frame() %>% 
  {
  upset(., ncol(.), 20, show.numbers = TRUE, order.by = "freq",
      main.bar.color = "#6E8691",
      matrix.color = "#6E8691",
      sets.bar.color = "#53AC9B")
  }
```

```{r}
exclusion_reasons_hc_table %>% 
  knitr::kable(caption = "Reasons for exclusion. _n_ shows the number of affected women, _unique_ those who for whom this was the only reason to be excluded.")
```


## In the diary
```{r}
exclusion_reasons_diary %>% 
  filter(included == 0) %>%
  select(-session, -included, -didnt_do_diary) %>% 
  as.data.frame() %>% 
  {
  upset(., ncol(.), 20, show.numbers = TRUE, order.by = "freq",
      main.bar.color = "#6E8691",
      matrix.color = "#6E8691",
      sets.bar.color = "#53AC9B")
  }

```

```{r}
exclusion_reasons_diary_table %>% 
  knitr::kable(caption = "Reasons for exclusion. _n_ shows the number of affected women, _unique_ those who for whom this was the only reason to be excluded.")
```

```{r}
included <- diary %>% filter(reasons_for_exclusion == "") %>% distinct(session)

testthat::expect_equal(
  all_surveys %>% filter(reasons_for_exclusion == "") %>% nrow(), 
  nrow(included)
)
```


# Sanity Checks
```{r}
library(testthat)
expect_equal(nrow(diary), diary_length)
expect_equal(nrow(all_surveys), all_survey_length)
expect_equal(nrow(diary_social), diary_social_length)
expect_false(any(names(diary) %contains% ".x"))
expect_false(any(names(diary) %contains% ".y"))
expect_false(any(names(all_surveys) %contains% ".y"))
expect_equal(groups(s3_daily), list())
expect_equal(groups(diary), list())
expect_equal(groups(all_surveys), list())
expect_equal(sum(duplicated(all_surveys$session)), 0)
expect_equal(sum(duplicated(s1_demo$session)), 0)
expect_equal(diary %>% drop_na(session, day_number) %>% 
               group_by(short, day_number) %>% filter(n() > 1) %>% nrow(), 0)
expect_equal(diary %>% drop_na(session, created_diary) %>%  
            group_by(session, created_diary) %>% filter(n()>1) %>% nrow(), 0)
expect_equal(s3_daily %>% drop_na(session, created_date) %>%  
            group_by(session, created_date) %>% filter(n()>1) %>% nrow(), 0)
expect_equal(diary %>% drop_na(session, created_date) %>%  
            group_by(session, created_date) %>% filter(n()>1) %>% nrow(), 0)
expect_equal(diary_social %>% drop_na(session, created_diary, person) %>%  
            group_by(session, created_diary, person) %>% filter(n() > 1) %>% nrow(), 0)
expect_equal(network %>% drop_na(session, person) %>%  
            group_by(session, person) %>% filter(n()>1) %>% nrow(), 0)
``` 


# Save
```{r}
save(diary_social, diary, sex_long, network_nominations, network, s1_demo, s1_filter, s2_initial, s3_daily, s4_followup, s4_timespent, withfollowup, all_surveys, file = "data/cleaned_selected.rdata")
```
